Cast these board IDs to ints to ensure no stringified numbers are entered

+++ system/ee/EllisLab/Addons/member/mod.member_auth.php
@@ -463,7 +463,7 @@ class Member_auth extends Member {
 				is_numeric(ee()->input->get_post('board_id')))
 			{
 				$query = ee()->db->select('board_label')
-					->where('board_id', ee()->input->get_post('board_id'))
+					->where('board_id', (int) ee()->input->get_post('board_id'))
 					->get('forum_boards');
 			}
 			else
@@ -574,7 +574,7 @@ class Member_auth extends Member {
 				is_numeric(ee()->input->get_post('board_id')))
 			{
 				$query = ee()->db->select("board_forum_url, board_label")
-					->where('board_id', ee()->input->get_post('board_id'))
+					->where('board_id', (int) ee()->input->get_post('board_id'))
 					->get('forum_boards');
 			}
 			else
@@ -689,7 +689,7 @@ class Member_auth extends Member {
 				is_numeric(ee()->input->get_post('board_id')))
 			{
 				$query = ee()->db->select('board_forum_url, board_id, board_label')
-					->where('board_id', ee()->input->get_post('board_id'))
+					->where('board_id', (int) ee()->input->get_post('board_id'))
 					->get('forum_boards');
 			}
 			else
